Move player info gui code out of the game file
